Update main.c
This is the working code with some operations converted to independent function operations. In addition, we should also write functions for all of the S and T keystream-generating operations and probably the decryption operation. Then we might add assertions(?). //NOT SURE HOW TO RUN THE CODE ON QUARK. THIS IS UNTESTED ON QUARK